run: change the type of max_retries_per_op from usize to u64
The value for this configuration option is parsed as u64 by the s-b
frontend, so changing this field's type from usize to u64 allows to get
rid of some casts.

+++ src/configuration.rs
@@ -35,7 +35,7 @@ pub struct Configuration {
 
     /// The maximum number of attempts an operation should be retried
     /// before giving up.
-    pub max_retries_per_op: usize,
+    pub max_retries_per_op: u64,
 }
 
 /// Contains all necessary context needed to execute an Operation.
